pyproject.toml
[project]
name = "finbrodataprep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "financedatabase>=2.3.1",
    "financetoolkit>=2.0.4",
    "ibapi>=9.81.1.post1",
    "exa-py>=0.0.1",
    "openai>=1.0.0",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "duckduckgo-search>=8.1.1",
    "tavily-python>=0.7.10",
    "pydantic-settings>=2.10.1",
    "finagg>=1.0.3",
    "yfinance>=0.2.28",
    "ipykernel>=6.30.0",
]
